Add files via upload
Conflict Resolution Algorithm in a Tangled Directed graphs or “two of every kind”.
Ten royal grooms want to choose brides from ten kingdoms as their wives, but at the same time the royal sympathies are ambiguous (everyone likes not one, but several princesses). How to unequivocally satisfy their expectations - so that each prince has a special princess and so that no one fights.
Fast and concise implementation of Constraint programming (CP) technology based on the free Google OR-Tools software.